chore(runtime): Rename file and add tests
